Add new cases to concurrent sessions coverage
