+++ drools-compiler/src/test/java/org/drools/compiler/integrationtests/facts/FactWithCharacter.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2017 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.drools.compiler.integrationtests.facts;
+
+public class FactWithCharacter {
+
+    private final char charValue;
+    private final Character characterValue;
+
+    public FactWithCharacter(final char charValue) {
+        this.charValue = charValue;
+        this.characterValue = charValue;
+    }
+
+    public char getCharValue() {
+        return charValue;
+    }
+
+    public Character getCharacterValue() {
+        return characterValue;
+    }
+}
