+++ drools-compiler/src/test/java/org/drools/compiler/integrationtests/facts/FactWithInteger.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2017 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.drools.compiler.integrationtests.facts;
+
+public class FactWithInteger {
+
+    private final int intValue;
+    private final Integer integerValue;
+
+    public FactWithInteger(final int intValue) {
+        this.intValue = intValue;
+        this.integerValue = intValue;
+    }
+
+    public int getIntValue() {
+        return intValue;
+    }
+
+    public Integer getIntegerValue() {
+        return integerValue;
+    }
+}
